Fix action trigger for circleci build event rules
The circleci build events from the release process does not have a ST2_COMMIT_SHA in the build parameters. The action execution of `st2ci.st2_pkg_e2e_test` being triggered are failing with error `Failed to render parameter "initial_commit"` because `has no attribute 'ST2_COMMIT_SHA'`. The value assignment of the `initial_commit` parameter is changed to the same if else Jinja expression for assigning hostname. If build_parameters are not present, then use the vcs_revision in the payload.