Get st2 version from installed package
If version is provided as input to the e2e test, use the given version. If version is not provided to the e2e test, get the version from the installed st2 package. If the installed st2 package is a dev version, leave the version null. Otherwise, use the installed version which will dictate which st2tests branch to install.